feat: Use EMA filter
This is "faster" and doesn't require an array of the moving average filter I
had before. I may have something wrong but the statement to only update
on "changes", which I want to do to minimize the print noise, causes the LED
to "never" turn off when the pot is turned off. Adding the diff_clamp_to_0
resolves the problem although there is probably a better way.

Also, adding the delay between samples seems to make things "smoother",
not sure why but feels that way to me.