feat: Add 21.1_Relay
With the circut using a relay is wired as provided in chapter 21.1 of the Freenove
FNK0045 Tutorial Relay.py turn the motor on when pressed and it's off when released.
This app also prints logging information via the serial port. This allows the program to
be run/tested without the circuit.